Revert "Refactor project name"
