+++ dependency-reduced-pom.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.trecboosters</groupId>
-  <artifactId>LuceneGroupProject</artifactId>
-  <name>LuceneGroupProject</name>
+  <groupId>com.assignment.lucene</groupId>
+  <artifactId>lucenmavenprojectnikhil</artifactId>
+  <name>lucenmavenprojectnikhil</name>
   <version>0.0.1-SNAPSHOT</version>
+  <url>http://www.example.com</url>
   <build>
     <plugins>
       <plugin>
