+++ pom.xml
@@ -4,11 +4,13 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.trecboosters</groupId>
-  <artifactId>LuceneGroupProject</artifactId>
+  <groupId>com.assignment.lucene</groupId>
+  <artifactId>lucenmavenprojectnikhil</artifactId>
   <version>0.0.1-SNAPSHOT</version>
 
-  <name>LuceneGroupProject</name>
+  <name>lucenmavenprojectnikhil</name>
+  <!-- FIXME change it to the project's website -->
+  <url>http://www.example.com</url>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
