+++ src/main/java/com/assignment/lucene/lucenmavenprojectnikhil/model/QueryModel.java
@@ -1,4 +1,4 @@
-package com.trecboosters.LuceneGroupProject.model;
+package com.assignment.lucene.lucenmavenprojectnikhil.model;
 
 public class QueryModel {
